+++ src/compas_timber/connections/t_birdsmouth.py
@@ -4,8 +4,8 @@ from compas.geometry import intersection_plane_plane_plane
 
 from compas_timber._fabrication import DoubleCut
 from compas_timber.connections.utilities import beam_ref_side_incidence
+from compas_timber.errors import BeamJoinningError
 
-from .joint import BeamJoinningError
 from .joint import Joint
 from .solver import JointTopology